fix(ci): keep models within line budget
